Fixed size_t conversion warnings. JIRA: CXX-5057
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@65029 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/data_loaders/genbank/processors.cpp
@@ -1879,7 +1879,7 @@ void CProcessor_ID2::x_ReadData(const CID2_Reply_Data& data,
     in->SetSkipUnknownMembers(eSerialSkipUnknown_Yes);
     in->SetSkipUnknownVariants(eSerialSkipUnknown_Yes);
     in->Read(object);
-    data_size += in->GetStreamPos();
+    data_size += size_t(in->GetStreamPos()); // in-memory size can't be bigger than size_t
 }
